Initial implementation of note map
Every block of music data is a time-to-notes dict. Each time index (key),
which represents the offset of events from the begin time of this block, is
mapped into note events of 2 kinds. A Note object indicates that it starts
at this time, and a weakref to a note is present when it continues at this
time.

[!] Note (Chord_Common_Data) and Music_Data_Block keep the absolute begin
    time from the begin of the score.
[!] Note events are sorted from low to high at the end of measure parsing.