Score: Integrate creation and indexing functions
The only method to access data in a score is __call__. The flag no_create
must be False to allow auto creation. If one of (part) id and (measure)
number is None, __call__ returns a part or a measure according to the non-
null parameter. If both parameters are present, it returns the specified
block of music data.